transfert_argent/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
idna==3.4
Markdown==3.4.4
psycopg2==2.9.7
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
